+++ packages/devtools_shared/CHANGELOG.md
@@ -2,7 +2,8 @@
 * Refactor yaml extension methods.
 * **Breaking change:** rename `ServerApi.getCompleted` to `ServerApi.success` and make the
 `value` parameter optional.
-* **Breaking change:** 
+* **Breaking change:** remove the `String dtdUri` parameter from `ServerApi.handle` and replace
+it with a record parameter `({String? uri, String? secret})? dtd`.
 
 # 7.0.0
 * **Breaking change:** remove the `ServerApi.setCompleted` method that was a